Исправления + улучшения
Поменял местами мутации, стало как будто хуже работать, хотя должно было лучше. Может не повезло и нужны ещё тесты
