+++ Data/Save_weights/best_weights.txt
@@ -1,16 +1,16 @@
-0.00966530
--0.75522799
--0.75874202
-0.00736067
-0.07576398
--1.00000000
--0.30840899
-0.06970972
-0.00677792
--0.54004792
--0.94229563
-0.07057999
-0.02719914
--0.01902863
--0.51243408
-0.13858485
+-0.30000000
+-0.63000000
+-0.69000000
+-0.63000000
+-0.70000000
+-0.90000000
+-0.52000000
+-0.40000000
+-0.40000000
+-0.59000000
+-0.74000000
+-0.38000000
+-0.30000000
+-0.10000000
+-0.53000000
+0.46000000
