+++ Data/Save_weights/checkpoint_weights.txt
@@ -1,16 +1,16 @@
-0.00968255
--0.81041118
--0.79997506
-0.00814426
-0.07790028
--1.00000000
--0.29102069
-0.06874136
-0.00000000
--0.50009563
--0.91388629
-0.06834982
-0.02719914
--0.45034119
--0.42911589
-0.14000000
+0.11000000
+-0.35000000
+-0.49000000
+0.01000000
+-0.64000000
+0.09000000
+0.06000000
+-0.59000000
+-0.42000000
+0.07000000
+0.06000000
+-0.45000000
+0.05000000
+-0.46000000
+-0.59000000
+-0.05000000
