+++ Data/Statistic/Metadata/2025-05-17_02-08-10.txt
@@ -1,0 +1,31 @@
+Информация об обучении
+
+Правильная картинка: [1 1 1 1 1 0 0 1 1 0 0 1 1 1 1 1]
+NET для активации: 0.4
+
+════════════════════════════════════════
+Популяция: 50
+Максимум эпох: 10000
+Итоговое количество эпох: 10000
+Коэффициент мутации: 0.3
+Время обучения: 609 минут 53 секунд
+
+Лучший результат:
+Фитнес: 0.99009901
+Веса:
+  Ген 1: 0.05000000
+  Ген 2: 0.01000000
+  Ген 3: 0.05000000
+  Ген 4: 0.01000000
+  Ген 5: 0.01000000
+  Ген 6: -0.68000000
+  Ген 7: -0.66000000
+  Ген 8: 0.05000000
+  Ген 9: 0.03000000
+  Ген 10: -0.90000000
+  Ген 11: -0.49000000
+  Ген 12: 0.06000000
+  Ген 13: 0.05000000
+  Ген 14: 0.01000000
+  Ген 15: 0.08000000
+  Ген 16: -0.01000000
