+++ Data/Statistic/Metadata/2025-05-17_12-42-22.txt
@@ -1,0 +1,31 @@
+Информация об обучении
+
+Правильная картинка: [0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1]
+NET для активации: 0.4
+
+════════════════════════════════════════
+Популяция: 50
+Максимум эпох: 10000
+Итоговое количество эпох: 34
+Коэффициент мутации: 0.3
+Время обучения: 5 минут 1 секунд
+
+Лучший результат:
+Фитнес: 1.00000000
+Веса:
+  Ген 1: -0.30000000
+  Ген 2: -0.63000000
+  Ген 3: -0.69000000
+  Ген 4: -0.63000000
+  Ген 5: -0.70000000
+  Ген 6: -0.90000000
+  Ген 7: -0.52000000
+  Ген 8: -0.40000000
+  Ген 9: -0.40000000
+  Ген 10: -0.59000000
+  Ген 11: -0.74000000
+  Ген 12: -0.38000000
+  Ген 13: -0.30000000
+  Ген 14: -0.10000000
+  Ген 15: -0.53000000
+  Ген 16: 0.46000000
