+++ Data/correct_img.py
@@ -1,7 +1,8 @@
 import numpy as np
 
 # Тут хранится единственно правильное изображение в 4 на 4 пикселя
-My_img = np.array([1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 0, 0, 1], dtype=np.uint8)
-# My_img = np.array([0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1], dtype=np.uint8)
+# My_img = np.array([1, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1], dtype=np.uint8) # Сложность высокая
+My_img = np.array([1, 0, 0, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1, 0, 0, 1], dtype=np.uint8) # Средний вариант, но 300 эпох должно хватать в среднем
+# My_img = np.array([0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1], dtype=np.uint8) # Самый лёгкий вариант и самый быстрый
 
 NET = 0.4
