+++ Genetic/Methods/Fit_func.py
@@ -1,5 +1,9 @@
 import numpy as np
 
+# Старая с пенальти. Не отражала реальных результатов. С пенальти показывает что уже около 1, а по факту лишь 16% верно распознаёт
+# Но оказалосб что для лёгких и средних значений лучше её использовать
+# Короче просто лучше не использовать сложные случаи искомого изображения
+
 # Кеш для уже вычисленных фитнесов
 fitness_cache = {}
 
@@ -52,3 +56,46 @@ def fitness_function_cached(weights):
     # Сохраняем в кеш
     fitness_cache[key] = fitness
     return fitness
+
+
+# import numpy as np
+#
+# # Кеш для уже вычисленных фитнесов
+# fitness_cache = {}
+#
+# # Импорт изображения и порога
+# from Data.correct_img import My_img, NET
+#
+# def fitness_function_cached(weights):
+#     """
+#     Строгая фитнес-функция:
+#     - 0.0, если правильное изображение не проходит порог
+#     - 1.0 / (количество ложных срабатываний + 1) в остальных случаях
+#     """
+#     key = tuple(np.round(weights, 5))  # округляем, чтобы убрать шум
+#     if key in fitness_cache:
+#         return fitness_cache[key]
+#
+#     weights = np.array(weights, dtype=np.float32)
+#
+#     # Проверка правильного изображения
+#     if np.dot(weights, My_img) < NET:
+#         fitness_cache[key] = 0.0
+#         return 0.0
+#
+#     # Генерация всех бинарных изображений
+#     correct_int = int(''.join(map(str, My_img)), 2)
+#     false_positives = 0
+#
+#     for i in range(2 ** 16):
+#         if i == correct_int:
+#             continue
+#         bits = np.array([int(b) for b in format(i, '016b')], dtype=np.uint8)
+#         score = np.dot(weights, bits)
+#         if score >= NET:
+#             false_positives += 1
+#
+#     fitness = 1.0 / (false_positives + 1)
+#     fitness_cache[key] = fitness
+#     return fitness
+#
